backend/requirements.txt
flask
flask CORS
tensorflow
numpy
pillow
werkzeug
gunicorn
